+++ src/main/java/refinedstorage/proxy/ClientProxy.java
@@ -31,7 +31,8 @@ public class ClientProxy extends CommonProxy {
             new ResourceLocation("refinedstorage:4k_storage_disk"),
             new ResourceLocation("refinedstorage:16k_storage_disk"),
             new ResourceLocation("refinedstorage:64k_storage_disk"),
-            new ResourceLocation("refinedstorage:creative_storage_disk")
+            new ResourceLocation("refinedstorage:creative_storage_disk"),
+            new ResourceLocation("refinedstorage:debug_storage_disk")
         );
 
         ModelBakery.registerItemVariants(RefinedStorageItems.STORAGE_PART,
@@ -69,6 +70,7 @@ public class ClientProxy extends CommonProxy {
         ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_DISK, ItemStorageDisk.TYPE_16K, new ModelResourceLocation("refinedstorage:16k_storage_disk", "inventory"));
         ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_DISK, ItemStorageDisk.TYPE_64K, new ModelResourceLocation("refinedstorage:64k_storage_disk", "inventory"));
         ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_DISK, ItemStorageDisk.TYPE_CREATIVE, new ModelResourceLocation("refinedstorage:creative_storage_disk", "inventory"));
+        ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_DISK, ItemStorageDisk.TYPE_DEBUG, new ModelResourceLocation("refinedstorage:debug_storage_disk", "inventory"));
 
         ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_PART, ItemStoragePart.TYPE_1K, new ModelResourceLocation("refinedstorage:1k_storage_part", "inventory"));
         ModelLoader.setCustomModelResourceLocation(RefinedStorageItems.STORAGE_PART, ItemStoragePart.TYPE_4K, new ModelResourceLocation("refinedstorage:4k_storage_part", "inventory"));